ve: fix ve task state percpu counters
Counters overlap detection for ve tasks in running/uninterraprible/iowait state
was broken due to type mismatch:
nr_{running/unin..e/iowait}_ve() uses _long_ for summing _int_ percpu counters.

As result, it broke ve loadavg calculation after first int overlap.

This patch expand all this percpu counters to unsigned long.

http://bugzilla.openvz.org/show_bug.cgi?id=1396

+++ include/linux/ve.h
@@ -129,9 +129,9 @@ struct ve_cpu_stats {
 	cycles_t	strt_idle_time;
 	cycles_t	used_time;
 	seqcount_t	stat_lock;
-	int		nr_running;
-	int		nr_unint;
-	int		nr_iowait;
+	unsigned long	nr_running;
+	unsigned long	nr_unint;
+	unsigned long	nr_iowait;
 	cputime64_t	user;
 	cputime64_t	nice;
 	cputime64_t	system;
